Refactor: RecommendService 리팩토링 ======================================================================
[리팩토링 사유]
서비스 디렉토리 가독성 문제 및 추상화 이슈

[변경 진행사항]
- 1차적으로 추천 서비스에 관련 코드를 병합하는 방향으로 수정
