--- src/main/java/co/pickcake/chatGPT/response/ChatGptUsages.java
@@ -1,4 +1,0 @@
-package co.pickcake.chatGPT.response;
-
-public class ChatGptUsages {
-}
