+++ src/main/java/co/pickcake/recommend/chatGPT/query/event/RecommencementSpecific.java
@@ -1,4 +1,4 @@
-package co.pickcake.chatGPT.query.event;
+package co.pickcake.recommend.chatGPT.query.event;
 
 import lombok.Builder;
 import lombok.Data;
