+++ src/main/java/co/pickcake/recommend/chatGPT/query/event/RecommendEventType.java
@@ -1,4 +1,4 @@
-package co.pickcake.chatGPT.query.event;
+package co.pickcake.recommend.chatGPT.query.event;
 
 public enum RecommendEventType {
     WEATHER,
