+++ src/main/java/co/pickcake/recommend/chatGPT/query/event/SpecialEvent.java
@@ -1,4 +1,4 @@
-package co.pickcake.chatGPT.query.event;
+package co.pickcake.recommend.chatGPT.query.event;
 
 public enum SpecialEvent {
     VALENTINE,
