+++ src/main/java/co/pickcake/recommend/chatGPT/query/event/WeatherEvent.java
@@ -1,4 +1,4 @@
-package co.pickcake.chatGPT.query.event;
+package co.pickcake.recommend.chatGPT.query.event;
 
 public enum WeatherEvent {
     /* 유지 보수때문에 string 으로 치환해야할 것 같음. 지금 버전에서는 수정하지말고 아래에만 추가하도록 */
