+++ src/main/java/co/pickcake/recommend/chatGPT/query/item/RecommendItem.java
@@ -1,4 +1,4 @@
-package co.pickcake.chatGPT.query.item;
+package co.pickcake.recommend.chatGPT.query.item;
