+++ src/main/java/co/pickcake/recommend/response/ChatGptUsages.java
@@ -1,0 +1,4 @@
+package co.pickcake.recommend.response;
+
+public class ChatGptUsages {
+}
