+++ src/main/java/co/pickcake/reservedomain/entity/Reservation.java
@@ -1,0 +1,17 @@
+package co.pickcake.reservedomain.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Reservation {
+    @Id
+    @GeneratedValue
+    private Long reserveId;
+
+}
