+++ src/main/java/co/pickcake/reservedomain/searchcake/repository/CakeRepository.java
@@ -1,7 +1,6 @@
 package co.pickcake.reservedomain.searchcake.repository;
 
 import co.pickcake.reservedomain.entity.item.Cake;
-import co.pickcake.reservedomain.searchcake.dto.CakeSearch;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -38,12 +37,5 @@ public interface CakeRepository extends JpaRepository<Cake, Long> {
             " join fetch c.cakeImages ci")
     Page<Cake> findAllByPaging(Pageable pageable);
 
-
-    /* TODO 동적 쿼리는 만들다 보니 현 형태로는 이슈가 있음, 추후에 QueryDSL 로 바꿔야야 함 */
-//    @Query("select c from Cake c" +
-//            " join fetch c.cakeImages ci" +
-//            " join fetch c.cakeCategoryList catl" +
-//            " where c.brand = :brand")
-//    Page<Cake> findBySpecific(CakeSearch cakeSearch, Pageable pageable);
 
 }
