+++ src/test/java/co/pickcake/chatGPT/service/ChatGPTServiceTest.java
@@ -1,20 +1,19 @@
 package co.pickcake.chatGPT.service;
 
-import co.pickcake.chatGPT.query.RecommendEvent;
-import co.pickcake.chatGPT.query.RecommendQuery;
-import co.pickcake.chatGPT.query.RecommendType;
-import co.pickcake.chatGPT.query.event.RecommencementSpecific;
-import co.pickcake.chatGPT.query.event.RecommendEventType;
-import co.pickcake.chatGPT.query.event.SpecialEvent;
-import co.pickcake.chatGPT.query.item.RecommendItem;
-import co.pickcake.chatGPT.response.ChatRecommendResponse;
+import co.pickcake.recommend.chatGPT.query.RecommendEvent;
+import co.pickcake.recommend.chatGPT.query.RecommendQuery;
+import co.pickcake.recommend.chatGPT.query.RecommendType;
+import co.pickcake.recommend.chatGPT.query.event.RecommencementSpecific;
+import co.pickcake.recommend.chatGPT.query.event.RecommendEventType;
+import co.pickcake.recommend.chatGPT.query.event.SpecialEvent;
+import co.pickcake.recommend.chatGPT.query.item.RecommendItem;
+import co.pickcake.recommend.response.ChatRecommendResponse;
+import co.pickcake.recommend.service.ChatGPTService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class ChatGPTServiceTest {
